fix: prevent server error when appending to Document
Prevent server error when appending a content to a Document, by checking
if the cursor is on the last position and changing the logic
accordingly.

Make the cursor follow the insertion of content into a Document, that
is, if a string content of length X is inserted at position Y, then the
current position will be set to X + Y.

Add specific configuration file for debugging in VS Code to the folder
".vscode".